requirements.txt
selenium==4.12.0
webdriver-manager==4.2.0
Python 3.14

pages
pytest
allure-python-commons
driver